Add files via upload
In this JavaScript program I created a system that tells the user if certain subjects were approved to pass the final scores. ideal for schools, universities and course platforms to evaluate their students's performance. Eventhough is a basic program that executes in console, It's very useful to understand the logic and it's prompt to be used with HTML and CSS